Extract forms to own module
